requirement.txt
asgiref==3.6.0
Django==4.1.1
django-autoslug==1.9.8
django-crispy-forms==2.0
django-decouple==2.1
django-dotenv==1.4.2
djangorestframework==3.14.0
Pillow==9.4.0
python-decouple==3.8
pytz==2023.3
sqlparse==0.4.3
tzdata==2022.7
